Cleaned up Flagged.tsx, and converted utils and testing files to TypeScript
