--- client/src/utils/utils.js
@@ -1,5 +1,0 @@
-// utility file that contains helper functions for the client
-
-export function demo_sum(a, b) {
-    return a + b
-}
